--- CookingWebsite/CookingWebsite/Modules/RecipeCreateModule/RecipeCreateBindings.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace CookingWebsite.Modules.RecipeCreateModule
-{
-    public static class RecipeCreateBindings
-    {
-        public static IServiceCollection AddRecipeCreateModule(this IServiceCollection services)
-        {
-            return services;
-        }
-    }
-}
