--- CookingWebsite/CookingWebsite/Modules/RecipeCreateModule/RecipeIngredientDto.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace CookingWebsite.Modules.RecipeCreateModule
-{
-    public class RecipeIngredientDto
-    {
-        public int RecipeId { get; set; }
-        public string Title { get; set; }
-        public List<RecipeIngredientItemDto> Items { get; set; }
-    }
-}
