+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeCreateUpdateBindings.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
+{
+    public static class RecipeCreateUpdateBindings
+    {
+        public static IServiceCollection AddRecipeCreateUpdateModule(this IServiceCollection services)
+        {
+            return services;
+        }
+    }
+}
