+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeCreateUpdateController.cs
@@ -9,23 +9,23 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 
-namespace CookingWebsite.Modules.RecipeCreateModule
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
 {
     [ApiController]
-    [Route("api/recipes/create")]
-    public class RecipeCreateController : ControllerBase
+    [Route("api/recipes")]
+    public class RecipeCreateUpdateController : ControllerBase
     {
 
         private readonly IRecipeService _recipeService;
         private readonly IUnitOfWork _unitOfWork;
 
-        public RecipeCreateController(IUnitOfWork unitOfWork, IRecipeService recipeService)
+        public RecipeCreateUpdateController(IUnitOfWork unitOfWork, IRecipeService recipeService)
         {
             _unitOfWork = unitOfWork;
             _recipeService = recipeService;
         }
 
-        [HttpPost]
+        [HttpPost("create")]
         public async Task AddRecipe()
         {
             IFormFileCollection files = Request.Form.Files;
@@ -35,6 +35,20 @@ namespace CookingWebsite.Modules.RecipeCreateModule
             var recipeDto = await addRecipeDto.Map(files, User.FindFirstValue(Claims.Username));
 
             await _recipeService.AddRecipe(recipeDto);
+
+            await _unitOfWork.Commit();
+        }
+
+        [HttpPost("{recipeId}/update")]
+        public async Task UpdateRecipe()
+        {
+            IFormFileCollection files = Request.Form.Files;
+
+            UpdateRecipeDto updateRecipeDto = JsonConvert.DeserializeObject<UpdateRecipeDto>(Request.Form["data"]);
+
+            var recipeDto = await updateRecipeDto.Map(files);
+
+            await _recipeService.UpdateRecipe(recipeDto);
 
             await _unitOfWork.Commit();
         }
