+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeCreateUpdateMapper.cs
@@ -3,10 +3,56 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CookingWebsite.Modules.RecipeUpdateModule
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
 {
-    public static class UpdateRecipeMapper
+    public static class RecipeCreateUpdateMapper
     {
+        public async static Task<Application.Recipes.RecipeDtos.AddRecipeDto> Map(
+            this AddRecipeDto recipe,
+            IFormFileCollection files,
+            string authorizedUser)
+        {
+            var result = new Application.Recipes.RecipeDtos.AddRecipeDto();
+
+            result.Title = recipe.Title;
+            result.Description = recipe.Description;
+            result.CookingTime = recipe.CookingTime;
+            result.PersonsCount = recipe.PersonsCount;
+
+            result.Ingredients = new List<Application.Recipes.RecipeDtos.RecipeIngredientDto>();
+
+            result.Ingredients = recipe.Ingredients.Select(x => new Application.Recipes.RecipeDtos.RecipeIngredientDto
+            {
+                Title = x.Title,
+                Items = x.Items.Select(y => new Application.Recipes.RecipeDtos.RecipeIngredientItemDto
+                {
+                    Name = y.Name,
+                    Value = y.Value
+                }).ToList()
+            }).ToList();
+
+            result.Steps = new List<Application.Recipes.RecipeDtos.RecipeStepDto>();
+
+            result.Steps = recipe.Steps.Select((s, i) => new Application.Recipes.RecipeDtos.RecipeStepDto
+            {
+                StepNumber = i + 1,
+                Description = s.Description
+            }).ToList();
+
+            result.Tags = new List<Application.Recipes.RecipeDtos.RecipeTagDto>();
+
+            result.Tags = recipe.Tags.Select(x => new Application.Recipes.RecipeDtos.RecipeTagDto
+            {
+                TagName = x
+            }).ToList();
+
+            result.AuthorUsername = authorizedUser;
+
+            result.Image = await FileManagment.CreateAsync(files[0]);
+
+            return result;
+        }
+
         public async static Task<Application.Recipes.RecipeDtos.UpdateRecipeDto> Map
         (
             this UpdateRecipeDto recipe,
