+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeIngredientDto.cs
@@ -1,6 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
-namespace CookingWebsite.Modules.RecipeUpdateModule
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
 {
     public class RecipeIngredientDto
     {
