+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeIngredientItemDto.cs
@@ -1,4 +1,9 @@
-﻿namespace CookingWebsite.Modules.RecipeUpdateModule
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
 {
     public class RecipeIngredientItemDto
     {
