+++ CookingWebsite/CookingWebsite/Modules/RecipeCreateUpdateModule/RecipeStepDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CookingWebsite.Modules.RecipeCreateModule
+namespace CookingWebsite.Modules.RecipeCreateUpdateModule
 {
     public class RecipeStepDto
     {
