--- CookingWebsite/CookingWebsite/Modules/RecipeUpdateModule/RecipeStepDto.cs
@@ -1,7 +1,0 @@
-﻿namespace CookingWebsite.Modules.RecipeUpdateModule
-{
-    public class RecipeStepDto
-    {
-        public string Description { get; set; }
-    }
-}
