--- CookingWebsite/CookingWebsite/Modules/RecipeUpdateModule/RecipeUpdateBindings.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace CookingWebsite.Modules.RecipeUpdateModule
-{
-    public static class RecipeUpdateBindings
-    {
-        public static IServiceCollection AddRecipeUpdateModule(this IServiceCollection services)
-        {
-            return services;
-        }
-    }
-}
